+++ rewrite_in_rust/src/main.rs
@@ -1,5 +1,12 @@
+mod expr;
 mod lexer;
+mod parser;
+mod state;
+mod stmt;
+mod tokens;
+mod tree_interp;
 mod vm;
+use fnv::FnvHashMap;
 
 use std::env;
 use std::fs::read_to_string;
@@ -11,6 +18,25 @@ while a < 10000000 {
     a = a + 1
 }
 ";
+
+fn tree_interpreter(source: String) {
+    let state = &mut state::InterpreterState::new(source);
+    let mut lex = lexer::Lexer::new(state);
+    lex.scan();
+    let mut parser = parser::Parser::new(state);
+    parser.parse();
+    // println!("{:#?}", state.tokens);
+    // println!("{:#?}", state.stmts);
+    let mut env = tree_interp::Env {
+        enclosing: None,
+        values: FnvHashMap::default(),
+    };
+    let mut exec = tree_interp::Exec::new(core::ptr::addr_of_mut!(env));
+    let start = Instant::now();
+    exec.interpret(&mut state.stmts);
+    let duration = start.elapsed();
+    println!("Duration tree: {:?}", duration.as_secs_f64());
+}
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -26,5 +52,5 @@ fn main() {
     vm::test_vm_execution();
     let duration = start.elapsed();
     println!("Duration bytecode: {:?}", duration.as_secs_f64());
-    lexer::scan(String::from(source));
+    tree_interpreter(String::from(source));
 }
