+++ rewrite_in_rust/src/state.rs
@@ -1,0 +1,37 @@
+use crate::stmt::*;
+use crate::tokens::*;
+
+#[derive(Debug)]
+pub struct InterpreterError {
+    pub message: String,
+    pub line: i32,
+    pub pos: usize,
+}
+
+#[derive(Debug)]
+pub struct InterpreterState {
+    pub source: String,
+    pub tokens: Vec<TokenData>,
+    pub errors: Vec<InterpreterError>,
+    pub stmts: Vec<Stmt>,
+}
+
+impl InterpreterState {
+    pub fn new(source: String) -> Self {
+        InterpreterState {
+            source: source,
+            tokens: vec![],
+            errors: vec![],
+            stmts: vec![],
+        }
+    }
+
+    pub fn set_error(&mut self, err: InterpreterError) {
+        self.errors.push(err);
+    }
+
+    pub fn fatal_error(&mut self, err: InterpreterError) {
+        self.errors.push(err);
+        panic!();
+    }
+}
